fix: 시스템 메시지 확인 추가 및 삭제 가능 메시지 조건 개선

+++ bot/Commands/utility/deleteMsg.js
@@ -12,16 +12,21 @@ async function delayedDeleteMessage(message, delay) {
 		message = await message.interaction.fetchReply();
 	}
 	setTimeout(() => {
-		message.delete()
-			.then(() => console.log("메시지가 성공적으로 삭제되었습니다."))
-			.catch(error => {
-				const errorMessage =
-					error.code === 50013
-						? "메시지를 삭제할 권한이 없습니다."
-						: `메시지를 삭제하는 도중 오류가 발생했습니다: ${error.message}`;
-				console.log(errorMessage);
-				// .catch(() => console.error("오류 메시지를 전송할 수 없습니다."));
-			});
+		// 시스템 메시지 확인 추가
+		if (message.deletable && !message.system) {
+			message.delete()
+				.then(() => console.log("메시지가 성공적으로 삭제되었습니다."))
+				.catch(error => {
+					const errorMessage =
+						error.code === 50013
+							? "메시지를 삭제할 권한이 없습니다."
+							: `메시지를 삭제하는 도중 오류가 발생했습니다: ${error.message}`;
+					console.log(errorMessage);
+					// .catch(() => console.error("오류 메시지를 전송할 수 없습니다."));
+				});
+		} else {
+			console.log("시스템 메시지이거나 삭제할 수 없는 메시지입니다.");
+		}
 	}, delay * 1000);
 }
